+++ 01.Core/Sheep.Core.Application/Sheep/PricePeriod/CreateCommand.cs
@@ -7,8 +7,8 @@ namespace Sheep.Core.Application.Sheep.PricePeriod
     public class CreateCommand
     {
         public Guid SheepNumber { get; set; }
-        public long? PriceSheep { get; set; }
-        public long? Unabsorbedcosts { get; set; }
+        public double? PriceSheep { get; set; }
+        public double? Unabsorbedcosts { get; set; }
         public double PricePerSheep { get; set; }
        public Guid CategoryPriceId { get; set; }
        public GenderType Gender { get; set; }
