+++ 01.Core/Sheep.Core.Application/Sheep/SheepFullPrice/CreateCommand.cs
@@ -3,8 +3,8 @@ namespace Sheep.Core.Application.Sheep.SheepFullPrice
 {
     public class CreateCommand
     {
-        public long? PriceSheep { get; set; }
-        public long? Unabsorbedcosts { get; set; }
+        public double? PriceSheep { get; set; }
+        public double? Unabsorbedcosts { get; set; }
         public Guid SheepId { get; set; }
         public DateTime Calcuted { get; set; }
     }
