+++ 01.Core/Sheep.Core.Domain/Sheep/Entities/SheepPricePeriodEntity.cs
@@ -8,15 +8,15 @@ namespace Sheep.Core.Domain.Sheep.Entities
     {
         public Guid SheepId { get; set; }
         public Guid CategoryPriceId { get; set; }
-        public long? PriceSheep { get; set; }
-        public long? Unabsorbedcosts { get; set; }
+        public double? PriceSheep { get; set; }
+        public double? Unabsorbedcosts { get; set; }
         public DateTime Start {  get; set; }
         public DateTime End { get; set; }
         public DateTime Calcuted { get; set; }
         public SheepEntity SheepEntity { get; set; }
         public CategoryPriceEntity CategoryPriceEntity { get; set; }
         public SheepPricePeriodEntity() { }
-        public SheepPricePeriodEntity(Guid sheepId, Guid categoryPriceId, long? pricesheep,long? unabsorbedcosts ,DateTime start,DateTime end, DateTime calcuted)
+        public SheepPricePeriodEntity(Guid sheepId, Guid categoryPriceId, double? pricesheep, double? unabsorbedcosts ,DateTime start,DateTime end, DateTime calcuted)
         {
             PriceSheep = pricesheep;
             Unabsorbedcosts = unabsorbedcosts;
@@ -26,7 +26,7 @@ namespace Sheep.Core.Domain.Sheep.Entities
             CategoryPriceId=categoryPriceId;
             Calcuted = calcuted;
         }
-        public void Edit(Guid sheepId, Guid categoryPriceId, long? pricesheep, long? unabsorbedcosts, DateTime start, DateTime end , DateTime calcuted)
+        public void Edit(Guid sheepId, Guid categoryPriceId, double? pricesheep, double? unabsorbedcosts, DateTime start, DateTime end , DateTime calcuted)
         {
             PriceSheep = pricesheep;
             Unabsorbedcosts = unabsorbedcosts;
